+++ apps/web/src/app-routes.tsx
@@ -7,5 +7,8 @@ import { DashboardPage2 } from './features/dashboard/pages/dashboard-page2'
 // routes
 export const LandingPageRouteId = app.router.bind({ path: '/', element: <LandingPage /> })
 export const DashboardRouteId = app.router.bind({ path: '/dashboard', element: <DashboardLayout /> })
-export const DashboardPage1RouteId = app.router.bind({ path: 'page1', Component: DashboardPage1 }, DashboardRouteId)
+export const DashboardPage1RouteId = app.router.bind(
+    { path: 'page1', index: true, Component: DashboardPage1 },
+    DashboardRouteId
+)
 export const DashboardPage2RouteId = app.router.bind({ path: 'page2', Component: DashboardPage2 }, DashboardRouteId)