+++ apps/web/src/main.tsx
@@ -3,7 +3,11 @@ import './index.css'
 import React from 'react'
 import ReactDOM from 'react-dom/client'
 import { AppRouter, VinessReactApp } from '@viness/react'
+
 import { app } from './app'
+
+import './app-routes'
+import './app-stores'
 
 const ErrorFallbackComponent = () => <div>error</div>
 const SuspenseFallbackComponent = () => <div></div>