+++ packages/viness-react/src/i18n.ts
@@ -1,6 +1,8 @@
 import i18n, { InitOptions } from 'i18next'
 
 import { initReactI18next } from 'react-i18next'
+import { IVinessAppConfig } from './app-config'
+import { createIdentifier } from './identifier'
 
 export type VinessI18nConfig<T = object> = InitOptions<T>
 
@@ -17,3 +19,13 @@ export function initI18n<T>(option: InitOptions<T>) {
 }
 
 export { type InitOptions }
+
+export const II18n = createIdentifier<I18n>('II18n')
+
+export class I18n {
+    constructor(@IVinessAppConfig appConfig: IVinessAppConfig) {
+        if (appConfig.i18n) {
+            initI18n(appConfig.i18n)
+        }
+    }
+}